requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
trimesh>=3.9.0
shap-e @ git+https://github.com/openai/shap-e.git
huggingface-hub>=0.16.0
tqdm>=4.65.0
pillow>=9.5.0
clip @ git+https://github.com/openai/CLIP.git
ftfy>=6.1.1
regex>=2023.0.0
scipy>=1.11.0 